+++ cnccode/business_classes/BUSalesOrder.inc.php
@@ -1263,6 +1263,19 @@ class BUSalesOrder extends Business
                         DBEOrdline::sequenceNo,
                         $sequenceNo
                     );
+                    // get new prices
+                    $dbeItem=new DBEItem($this);
+                    $dbeItem->getRow( $dsNewOrdline->getValue(DBEOrdline::itemID));
+                    if($dbeItem->rowCount>0)
+                    {
+                        $curUnitCost=$dbeItem->getValue(DBEItem::curUnitCost);
+                        $curUnitSale=$dbeItem->getValue(DBEItem::curUnitSale);
+                        $qty= $dsNewOrdline->getValue(DBEOrdline::qtyOrdered);
+                        $dsNewOrdline->setValue(DBEOrdline::curUnitCost,$curUnitCost);
+                        $dsNewOrdline->setValue(DBEOrdline::curUnitSale,$curUnitSale);
+                        $dsNewOrdline->setValue(DBEOrdline::curTotalCost,round($qty*$curUnitCost,2));
+                        $dsNewOrdline->setValue(DBEOrdline::curTotalSale,round($qty*$curUnitSale,2));
+                    }
                     $dsNewOrdline->post();
                 }
             }